backend/requirements.txt
openai>=1.12.0
google-genai>=1.0.0
groq>=0.4.1
reportlab>=4.1.0
pydantic>=2.6.0
Pillow>=10.2.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
httpx>=0.27.0
deepgram-sdk>=3.0
